requirements.txt
BeautifulSoup==3.2.1
Flask==0.9
Jinja2==2.6
Pillow==2.4.0
Werkzeug==0.8.3
argparse==1.1
beautifulsoup4==4.3.2
pyjade==1.9
python-dateutil==1.5
python-firebase==1.2
requests==1.2.3
tornado==3.2
speechrecognition
nltk
numpy
wave
multiprocessing
pydub
